pyproject.toml
[project]
name = "mcp-gitlab"
version = "0.1.0"
description = "MCP server for GitLab On-Premise integration"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "python-gitlab>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-gitlab = "mcp_gitlab.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_gitlab"]
